requirements/base.txt


# Bleeding edge Django
django==1.9.6

# Configuration
django-environ==0.4.0
django-secure==1.0.1
whitenoise==3.0


# Forms
django-braces==1.8.1
django-crispy-forms==1.6.0
django-floppyforms==1.6.2

# Models
django-model-utils==2.5

# Images
Pillow==3.2.0

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.25.2


# Python-PostgreSQL Database Adapter
psycopg2==2.6.1

# Unicode slugification
unicode-slugify==0.1.3
django-autoslug==1.9.3

# Time zones support
pytz==2016.4

# Redis support
django-redis==4.4.2
redis>=2.10.0

celery==4.0.2

django-datatables-view==1.13.0
django-mptt==0.8.4
django-paintstore==0.2
django-makemessages-xgettext==0.1.1
djangorestframework==3.3.3  # move back 3.5.3
markdown==2.6.6
django-filter==0.13.0
django-rest-swagger==0.3.5  #move back 2.1.1
drf-nested-routers==0.11.1
xlsxwriter==0.9.2
tablib==0.11.2
django-import-export==0.5.0
django-suit==0.2.21
django-autocomplete-light==3.2.1
django-helpdesk==0.1.18
mongoengine==0.11.0
django-admin-rangefilter==0.3.0
django-bootstrap3==8.2.3
bootstrap3-datetime==2.4

# Your custom requirements go here
